+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## Version 5.4.5
+
+🐛 Bug Fix
+
+- Fixed encounter list for Safari Zone in Yellow
+- Fixed encounter list for Seafoam island in Fire Red and Leaf Green
+
+🏠 Internal
+
+- Bumped @sentry/react to 7.57.0
+- Bumped @sentry/tracing to 7.57.0
+- Bumped i18next to 23.2.6
+- Bumped react-i18next to 13.0.1
+- Bumped react-router-dom to 6.14.1
+- Bumped @cypress/code-coverage to 3.10.8
+- Bumped @types/jest to 29.5.2
+- Bumped @types/node to 20.3.3
+- Bumped @types/react to 18.2.14
+- Bumped @types/react-dom to 18.2.6
+- Bumped cypress to 12.16.0
+- Bumped sass to 1.63.6
+- Bumped sass-loader to 13.3.2
+
 ## Version 5.4.4
 
 🏠 Internal